+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Tao 2026 levofloxacin ([doi:10.1128/aac.01853-25](https://doi.org/10.1128/aac.01853-25)) -- Chinese children aged 0.16-16 years with severe refractory *Mycoplasma pneumoniae* pneumonia.
+- Add Lee 2026 oxacillin ([doi:10.1128/aac.01777-25](https://doi.org/10.1128/aac.01777-25)) -- preterm and term neonates and infants up to 90 days old on continuous infusion.
 
 - Add Padavia 2024 paracetamol and metabolites ([doi:10.1007/s40262-024-01439-3](https://doi.org/10.1007/s40262-024-01439-3)) -- extreme preterm neonates of 23-26 weeks' gestational age.